+++ db/migrate/20161111174946_remove_deprecated_grid_variables_from_variables.rb
@@ -1,0 +1,5 @@
+class RemoveDeprecatedGridVariablesFromVariables < ActiveRecord::Migration[5.0]
+  def change
+    remove_column :variables, :deprecated_grid_variables, :text
+  end
+end