+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 0.46.0
 
+### Refactoring
+- Removed deprecated_grid_variables from variables
+
 ## 0.45.0 (November 11, 2016)
 
 ### Enhancements